requirements.txt
# ── PDF parsing ───────────────────────────────────────────────────────────────
PyMuPDF>=1.24.0
pytesseract>=0.3.10   # OCR fallback for scanned PDFs (requires Tesseract binary)
Pillow>=10.0.0        # image rendering for OCR

# ── Embeddings & ML ───────────────────────────────────────────────────────────
openai>=1.30.0
torch>=2.0.0
sentence-transformers>=3.0.0

# ── Vector store ──────────────────────────────────────────────────────────────
qdrant-client>=1.9.0

# ── BM25 ──────────────────────────────────────────────────────────────────────
rank-bm25>=0.2.2
numpy>=1.26.0

# ── Reranker ──────────────────────────────────────────────────────────────────
cohere>=5.0.0         # optional: only if reranking.provider = "cohere"

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ── UI ────────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
pandas>=2.0.0
requests>=2.31.0

# ── Config ────────────────────────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0
